Fixed bug where checkboxes were not aligned and deleted repeated functions in forms files
