+++ backend/helpers/formsHelpers.js
@@ -102,7 +102,7 @@ function generateProjectForm(projectCount){
   return projectRequest
 }
 
-async function generateForms(effort, belbin, project, projectCount, unitId) {
+async function generateForms(belbin, effort, project, projectCount, unitId) {
 
     if (effort) {
         var effortFormBody = effortForm();
